Create Coffee central
Problem description:  Is it just a fad or is it here to stay? You’re not sure, but the steadily increasing number of coffee shops that are opening in your hometown has certainly become quite a draw. Apparently, people have become so addicted to coffee that apartments that are close to many coffee shops will actually fetch higher rents. This has come to the attention of a local real-estate company. They are interested in identifying the most valuable locations in the city in terms of their proximity to large numbers of coffee shops. They have given you a map of the city, marked with the locations of coffee shops. Assuming that the average person is willing to walk only a fixed number of blocks for their morning coffee, you have to find the location from which one can reach the largest number of coffee shops. As you are probably aware, your hometown is built on a square grid layout, with blocks aligned on north-south and east-west axes. Since you have to walk along streets, the distance between intersections (a, b) and (c, d) is |a − c| + |b− d|

Input

The input contains several test cases. Each test case describes a city. The first line of each test case contains four integers dx, dy, n, and q. These are the dimensions of the city grid dx × dy (1 ≤ dx, dy ≤ 1000), the number of coffee shops n (0 ≤ n ≤ 5 · 10^5), and the number of queries q (1 ≤ q ≤ 20). Each of the next n lines contains two integers xi and yi (1 ≤ xi ≤ dx, 1 ≤ yi ≤ dy); these specify the location of the i-th coffee shop. There will be at most one coffee shop per intersection. Each of the next q lines contains a single integer m (0 ≤ m ≤ 10^6), the maximal distance that a person is willing to walk for a cup of coffee. The last test case is followed by a line containing four zeros. 

Output

For each test case in the input,display case number Then display one line per query in the test case. Each line displays the maximum number of coffee shops reachable for the given query distance m followed by the optimal location. For example, the sample output shows that 3 coffee shops are withinquery distance 1 of the optimal location (3, 4), 4 shops are within query distance 2 of optimal location (2, 2), and 5 shops are within query distance 4 of optimal location (3, 1). If there are multiple optimal locations, pick the location that is furthest south (minimal positive integer y-coordinate). If there is still a tie, pick the location furthest west (minimal positive integer x-coordinate). Follow the format of the sample output.

Sample input
4 4 5 3
1 1
1 2
3 3
4 4
2 4
1
2
4
0 0 0 0
Sample output
Case 1:
3 (3,4)
4 (2,2)
5 (3,1)

APPROACH :

1. Fast scanner class to create object for scanning user inputs.

2. I the case of returning boolean value false for dx == dy && dy == n && n == q && q == 0 continue the loop.

4. Display case number and increment testcase value

5. Create and update the array using returned element from transformTo methods

6. Create a method to assign values as  max <-- tmp;  maxA = a; maxB = b; when tmp greater than max

7. Print the max value

8. If only row-1 is positive :

 sum <-- sum+ above where above <--prefixSums[row - 1][col]

If only col-1 is positive

left <-- prefixSums[row][col - 1];

 sum <--left;

If both are positive 

 aboveLeft <-- prefixSums[row - 1][col - 1];

 sum <-- sum- aboveLeft;

9.   update array with "sum" value

               prefixSums[row][col] <-- sum

10. Class 2 operations :

    transformTo() : Sets argument value as (x + y, x - y + 1000) 

    transformBack() : Sets argument value as (-x - (y - 1000)) / -2, (-x + (y - 1000)) / -2)

11. print maximum number of coffee shops reachable for the given query distance m followed by the optimal location.